requirements.txt
decorator==4.4.2
dill==0.3.1.1
-e git+git@github.com:shahcompbio/microscope_image_converter.git@65f1128182a9925d7085926ee11b73136dfb3953#egg=microscope_image_converter
networkx==2.4
numpy==1.18.2
pandas==1.0.3
-e git+https://github.com/shahcompbio/pypeliner.git@c05fe838d3d7aa60fc1a427b6d0c38eba9001eee#egg=pypeliner
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
